Route Groupsでサイドバーとコンテナー部分を共通化
