+++ src/app/(sidebar)/layout.tsx
@@ -1,0 +1,40 @@
+import "../styles/globals.css"
+import "../styles/pages.scss"
+
+import { Inter } from "next/font/google"
+
+import Container from "@/app/components/Container"
+
+import FooterMenu from "../components/FooterMenu"
+import Sidebar from "../components/Sidebar"
+
+const inter = Inter({ subsets: ["latin"] })
+
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+}
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <>
+      <html lang="en">
+        <body className={inter.className}>
+          <div className="flex w-full justify-center">
+            <Sidebar />
+            <div className=" w-full max-w-[480px]">
+              <Container>
+                {children}
+                <FooterMenu />
+              </Container>
+            </div>
+          </div>
+        </body>
+      </html>
+    </>
+  )
+}
